// src/app/generated/model/importRequestDTOv1.ts
/**
 * tkit-portal-server API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 4.4.0-SNAPSHOT
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ImportRequestDTOv1Portal } from './importRequestDTOv1Portal'
import { ImportRequestDTOv1ThemeImportData } from './importRequestDTOv1ThemeImportData'
import { MenuItemStructureDTOv1 } from './menuItemStructureDTOv1'

export interface ImportRequestDTOv1 {
  /**
   * Whether menu permissions should be synced to APM after import
   */
  synchronizePermissions?: boolean
  portal: ImportRequestDTOv1Portal
  /**
   * Menu Items structure imported with a portal
   */
  menuItems?: Array<MenuItemStructureDTOv1>
  themeImportData?: ImportRequestDTOv1ThemeImportData
}