// src/app/generated/model/menuItemStructureDTO.ts
/**
 * tkit-portal-server API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 4.4.0-SNAPSHOT
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface MenuItemStructureDTO {
  version?: number
  creationDate?: string
  creationUser?: string
  modificationDate?: string
  modificationUser?: string
  id?: string
  key?: string
  name?: string
  description?: string
  url?: string
  portalName?: string
  applicationId?: string
  disabled?: boolean
  position?: number
  badge?: string
  portalExit?: boolean
  parentItemId?: string
  i18n?: { [key: string]: string }
  children?: Array<MenuItemStructureDTO>
  roles?: Array<string>
}